requirements.txt
requests==2.32.3
PyYAML==6.0.2
pytest==8.3.5
xmltodict==0.14.2
locust==2.37.5



